textos de indicação colocados

+++ src/services/project.service.js
@@ -1,8 +1,11 @@
 import Project from '../models/Project.js';
 
+
 const findAll = () => Project.find();
 
 const findById = (id) => Project.findById(id);
+
+const findByUserId = (userId) => Project.find({ user: userId });
 
 const create = (body) => Project.create(body);
 
@@ -13,4 +16,4 @@ const update = (id, title, urlGithub, description, projectImage, tags) => Projec
 
 const deleteById = (id) => Project.findByIdAndDelete(id);
 
-export default { findAll, findById, create, update, deleteById }
+export default { findAll, findById, create, update, deleteById, findByUserId }
